uh-qemu-*: Add rng-tools as machine essential
The use of emulation inside a Docker container causes a problem with
entropy and, as consequence, the HTTPS handshake fails.

To shorten the time to acquire enough entropy, the rng-tools utility
can be used, speeding up the runtime tests.

The change has been done for all emulated platforms, so it is applied
to:

 - uh-qemu-armv5e
 - uh-qemu-i386
 - uh-qemu-x86-64

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>